--- README.md
@@ -1,3 +1,0 @@
-# Tar Editor
-
-A JavaScript text editor.
