+++ package.json
@@ -1,0 +1,17 @@
+{
+  "name": "tareditor",
+  "version": "1.2.3",
+  "description": "Tar Editor is a JS text editor.",
+  "main": "index.js",
+  "scripts": {
+    "build": "webpack",
+    "live": "webpack --watch"
+  },
+  "license": "MPL-2.0",
+  "dependencies": {
+    "webpack": "^5.11.0"
+  },
+  "devDependencies": {
+    "webpack-cli": "^4.2.0"
+  }
+}
