--- test.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<meta charset = "utf-8"/>
-	<title>Testing Tar Editor</title>
-	<style>
-		html, body {
-			width: 100%;
-			height: 100%;
-			margin: 0px;
-		}
-		.tar-wrap {
-			height: 16rem;
-			max-height: 16rem;
-			background-color: red;
-		}
-		.tar-scroller {
-			background-color: green;
-		}
-		.tar-textarea {
-			background-color: blue;
-		}
-	</style>
-</head>
-<body>
-	<main></main>
-	<script src = "./core/index.js"></script>
-	<script>
-		tar.init(document.querySelector("main")).setLN(true);
-	</script>
-</body>
-</html>
